+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+
+* [BREAKING] `builder` option in FutureComboBox was renamed to `itemBuilder`.
+* [BREAKING] Removed `.chip` factory from `VitComboBox`.
+* Added `labelStyle` option to combo box.
+* Added `overlayDecorationBuilder` option to customize the overlay container.
+
 ## 1.0.0
 
 * First release.